Sync: Security with Develop

+++ src/main/java/imdl/scalator/service/LevitaService.java
@@ -1,16 +1,18 @@
 package imdl.scalator.service;
 
+import imdl.scalator.controller.filter.LevitaFilter;
+import imdl.scalator.domain.LevitaResumed;
 import imdl.scalator.domain.exception.EntityNotFoundException;
 import imdl.scalator.domain.Instrumento;
 import imdl.scalator.domain.Levita;
 import imdl.scalator.domain.exception.RogueException;
 import imdl.scalator.domain.input.LevitaInput;
+import imdl.scalator.persistence.EscalaRepository;
 import imdl.scalator.persistence.LevitaRepository;
 import imdl.scalator.service.mapper.LevitaMapper;
 
-import java.util.Comparator;
-import java.util.List;
-import java.util.UUID;
+import java.time.LocalDate;
+import java.util.*;
 
 public class LevitaService {
 
@@ -18,19 +20,35 @@ public class LevitaService {
 
     private final InstrumentoService instrumentoService;
 
-    public LevitaService(LevitaRepository levitaRepository, InstrumentoService instrumentoService) {
+    private final EscalaRepository escalaService;
+
+    public LevitaService(LevitaRepository levitaRepository, InstrumentoService instrumentoService, EscalaRepository escalaService) {
         this.levitaRepository = levitaRepository;
         this.instrumentoService = instrumentoService;
+        this.escalaService = escalaService;
     }
 
-    public List<Levita> findAll(){
-        return levitaRepository.findAll().stream()
+    public List<Levita> findAll(LevitaFilter filter){
+        return levitaRepository.findAll(filter.getNome(), filter.getInstrumento()).stream()
                 .map(LevitaMapper::entityToDomain).sorted(Comparator.comparing(Levita::getNome)).toList();
     }
+    public List<LevitaResumed> findAllResumed(){
+        return levitaRepository.findAllResumed().stream().map(LevitaMapper::entityToDomainResumed)
+                .sorted(Comparator.comparing(LevitaResumed::getNome)).toList();
+    }
+
     public List<Levita> findAllById(List<UUID> ids){
         return levitaRepository.findAllById(ids).stream().map(LevitaMapper::entityToDomain).toList();
     }
 
+    public List<Levita> findAllDisponivel(LocalDate date){
+        return levitaRepository.findAll().stream().map(LevitaMapper::entityToDomain)
+                .filter(levita -> {
+                    if(levita.getAgenda() == null)
+                        return true;
+                    return !levita.getAgenda().contains(date);
+                }).sorted(Comparator.comparing(Levita::getNome)).toList();
+    }
     public List<Levita> findAllByInstrument(Long instrumento){
         return levitaRepository.findAllByInstrumento(instrumento).stream()
                 .map(LevitaMapper::entityToDomain).sorted(Comparator.comparing(Levita::getNome)).toList();
@@ -43,62 +61,104 @@ public class LevitaService {
     public Levita create(LevitaInput input){
         validateInput(input);
         Levita levita = inputToDomain(input);
+        levita.setAgenda(new ArrayList<>());
         return LevitaMapper.entityToDomain(levitaRepository.save(LevitaMapper.domainToEntity(levita)));
     }
-    public Levita update(UUID id, LevitaInput input){
-        Levita levita = LevitaMapper.entityToDomain(levitaRepository.findById(id)
+    public Levita update(LevitaInput input){
+        Levita levita = LevitaMapper.entityToDomain(levitaRepository.findById(input.getId())
                 .orElseThrow(() -> new EntityNotFoundException("Levita não encontrada.")));
         if(input.getNome() != null)
-            if(input.getNome().isBlank())
+            if (input.getNome().isBlank())
                 throw new RogueException("O nome está vazio.");
-            else
-                levita.setNome(input.getNome());
+            else levita.setNome(input.getNome());
         if(input.getInstrumentos() != null)
             levita.setInstrumentos(input.getInstrumentos().stream().map(instrumentoService::findById).toList());
         if(input.getContato() != null)
             levita.setContato(input.getContato());
         if(input.getEmail() != null)
             levita.setEmail(input.getEmail());
-        levitaRepository.save(LevitaMapper.domainToEntity(levita));
-        return levita;
+        if(input.getDescricao() != null)
+            levita.setDescricao(input.getDescricao());
+        return LevitaMapper.entityToDomain(levitaRepository.save(LevitaMapper.domainToEntity(levita)));
     }
 
     public void deleteLevita(UUID id){
+        if(escalaService.existsByLevita(id))
+            throw new RogueException("Levita está em uma escala.");
         levitaRepository.delete(levitaRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("Levita não encontrada.")));
     }
 
-    public Levita addInstrumento(UUID id, Long instrumento){
+    public Levita addInstrumento(UUID id, Long codInstrumento){ // TODO FIX
         Levita levita = LevitaMapper.entityToDomain(levitaRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("Levita não encontrada.")));
         List<Instrumento> instrumentos = levita.getInstrumentos();
-        instrumentos.add(instrumentoService.findById(instrumento));
-        levita.setInstrumentos(instrumentos);
-        return levita;
+        if(instrumentos.stream().map(Instrumento::getId).toList().contains(codInstrumento))
+            throw new RogueException("Levita já possui este instrumento.");
+        Instrumento instrumento = instrumentoService.findById(codInstrumento);
+        List<Instrumento> newInstrumentos = new ArrayList<>();
+        for (Instrumento value : instrumentos) {
+            if (instrumento.getId() < value.getId())
+                newInstrumentos.add(instrumento);
+            else
+                newInstrumentos.add(value);
+        }
+        levita.setInstrumentos(newInstrumentos);
+        return LevitaMapper.entityToDomain(levitaRepository.save(LevitaMapper.domainToEntity(levita)));
     }
 
     public Levita removeInstrumento(UUID id, Long instrumento){
         Levita levita = LevitaMapper.entityToDomain(levitaRepository.findById(id)
                 .orElseThrow(() -> new EntityNotFoundException("Levita não encontrada.")));
         List<Instrumento> instrumentos = levita.getInstrumentos();
-        if(!instrumentos.remove(instrumentoService.findById(instrumento)))
-            throw new EntityNotFoundException("Levita já não tocava tal instrumento.");
-        levita.setInstrumentos(instrumentos);
+        if(!instrumentos.stream().map(Instrumento::getId).toList().contains(instrumento))
+            throw new EntityNotFoundException("Levita já não possui este instrumento.");
+
+        List<Instrumento> newInstrumentos = new ArrayList<>();
+        instrumentos.forEach(i -> {
+            if(!Objects.equals(i.getId(), instrumento))
+                newInstrumentos.add(i);
+        });
+
+        levita.setInstrumentos(newInstrumentos);
+        return LevitaMapper.entityToDomain(levitaRepository.save(LevitaMapper.domainToEntity(levita)));
+    }
+
+    public List<LocalDate> getLevitaAgenda(UUID id){
+        return levitaRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Levita não encontrado.")).getAgenda();
+    }
+
+    public Levita updateAgentaFromALevita(UUID id){
+        Levita levita = LevitaMapper.entityToDomain(levitaRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Levita não encontrada.")));
+
+        List<LocalDate> agenda = levita.getAgenda();
+        agenda.removeIf(date -> date.isBefore(LocalDate.now().minusDays(30)));
+        levita.setAgenda(agenda);
+
+        levitaRepository.save(LevitaMapper.domainToEntity(levita));
         return levita;
     }
 
-    public Levita changeDisponivel(UUID id){
+    public Levita addDataInAgenda(UUID id, LocalDate date){
         Levita levita = LevitaMapper.entityToDomain(levitaRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFoundException("Levita não encontrada.")));
-        levita.setDisponivel(!levita.isDisponivel());
-        levitaRepository.save(LevitaMapper.domainToEntity(levita));
-        return levita;
+                .orElseThrow(() -> new EntityNotFoundException("Levita não encontrado.")));
+        List<LocalDate> agenda = new ArrayList<>();
+        if(levita.getAgenda() != null)
+            agenda = levita.getAgenda();
+        if(agenda.contains(date))
+            throw new RogueException("Data já inserida.");
+        if(date.isBefore(LocalDate.now()))
+            throw new RogueException("Data já passou.");
+        agenda.add(date);
+        levita.setAgenda(agenda);
+        return LevitaMapper.entityToDomain(levitaRepository.save(LevitaMapper.domainToEntity(levita)));
     }
 
     private void validateInput(LevitaInput input){
         if(input.getNome() == null || input.getNome().isBlank())
             throw new RogueException("O nome está vazio.");
-        if(input.getInstrumentos() == null)
+        if(input.getInstrumentos() == null || input.getInstrumentos().isEmpty())
             throw new RogueException("O instrumento está vazio.");
         if((input.getContato() == null||input.getContato().isBlank())
                 && (input.getEmail() == null ||input.getEmail().isBlank()))
@@ -107,11 +167,13 @@ public class LevitaService {
 
     private Levita inputToDomain(LevitaInput input){
         Levita levita = new Levita();
+        if (input.getId() != null)
+            levita.setId(input.getId());
         levita.setNome(input.getNome());
         levita.setInstrumentos(input.getInstrumentos().stream().map(instrumentoService::findById).toList());
         levita.setContato(input.getContato());
         levita.setEmail(input.getEmail());
-        levita.setDisponivel(input.isDisponivel());
+        levita.setDescricao(input.getDescricao());
         return levita;
     }
 }
